% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beh_mat.R
\name{beh_mat}
\alias{beh_mat}
\title{beh_mat melt function
Takes a matlab behmat file and converts it to long format
plus it assigns binary state columns to each behavioral state class}
\usage{
beh_mat(n)
}
\arguments{
\item{n}{number of time frames to bin. generally 400 for 20 minute video is standard}
}
\description{
beh_mat melt function
Takes a matlab behmat file and converts it to long format
plus it assigns binary state columns to each behavioral state class
}
\examples{
beh_mat()
}
