% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beh_mat_tidy.R
\name{beh_mat_tidy}
\alias{beh_mat_tidy}
\title{beh_mat_tidy
Takes a matlab behmat file and converts it to long format
plus it assigns binary state columns to each behavioral state class}
\usage{
beh_mat_tidy(bins, frames, interactive = TRUE, folder = folder)
}
\arguments{
\item{bins}{number of time frames to bin. generally 400 for 20 minute video is standard}

\item{frames}{number of frames in the video}
}
\description{
beh_mat_tidy
Takes a matlab behmat file and converts it to long format
plus it assigns binary state columns to each behavioral state class
}
\examples{
beh_mat_tidy
}
