% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_mat.R
\name{dir_mat}
\alias{dir_mat}
\title{dir_mat melt function
Takes a matlab dirmat file and converts it to long format}
\usage{
dir_mat(n)
}
\arguments{
\item{n}{number of time frames to bin. generally 400 for 20 minute video is standard}
}
\description{
dir_mat melt function
Takes a matlab dirmat file and converts it to long format
}
\examples{
dir_mat()
}

