% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_mat_tidy.R
\name{dir_mat_tidy}
\alias{dir_mat_tidy}
\title{dir_mat_tidy melt function
Takes a matlab-derived dirmat .csv file and converts it to long format}
\usage{
dir_mat_tidy(interactive = TRUE)
}
\description{
dir_mat_tidy melt function
Takes a matlab-derived dirmat .csv file and converts it to long format
}
\examples{
dir_mat_tidy()
}
