% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_reversals.R
\name{fix_reversals}
\alias{fix_reversals}
\title{fix_reversals function
attempts to fix mis-ID'ed reversals and carry forward any directional information.
For example, it uses previous direction to assign this value for pirouettes and pauses
requires plyr, zoo, dplyr and some data.table}
\usage{
fix_reversals(df)
}
\arguments{
\item{df}{is the input data frame > should be output from beh_merge_all() ie df or lingrad_mirror()}
}
\description{
fix_reversals function
attempts to fix mis-ID'ed reversals and carry forward any directional information.
For example, it uses previous direction to assign this value for pirouettes and pauses
requires plyr, zoo, dplyr and some data.table
}
\examples{
tot.mat<-fix_reversals(df)
}

