% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAmpVdelF.R
\name{getAmpVdelF}
\alias{getAmpVdelF}
\title{getAmpVdelF}
\usage{
getAmpVdelF(FileFilter, ...)
}
\arguments{
\item{FileFilter}{string to search/subset filenames}
}
\description{
getAmpVdelF
}
\examples{
data <- getAmpVdelF(FileFilter = N2)

}
