% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetStripeProp.R
\name{GetSripeProp}
\alias{GetSripeProp}
\title{using all_track_data file, determines whether a worm is in stripe (1) or out (0)
based on an ideal, half-width stripe}
\usage{
GetSripeProp(data)
}
\arguments{
\item{data}{input data of type all_track_data from Import.WL.data()}
}
\description{
using all_track_data file, determines whether a worm is in stripe (1) or out (0)
based on an ideal, half-width stripe
}
\examples{
data \%>\% GetStripeProp() -> data
}
