% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.WL.data.R
\name{Import.WL.data}
\alias{Import.WL.data}
\title{Import Wormlab data, merge and convert to long format with curve angle.
plus it assigns binary state columns to each behavioral state class -work in progress-}
\usage{
Import.WL.data(
  bin.length = 4,
  frame.rate,
  num.tracks,
  multiple,
  head.correct = FALSE,
  ...
)
}
\arguments{
\item{bin.length}{length of time bins in seconds. Used for state analysis}

\item{frame.rate}{video frame rate}

\item{num.tracks}{optional argument to limit input to certain number of worm tracks}
}
\description{
Import Wormlab data, merge and convert to long format with curve angle.
plus it assigns binary state columns to each behavioral state class -work in progress-
}
\examples{
Import.WL.data(bin.length = 2, frame.rate = 6)
}
