% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_negdelta.R
\name{max_negdelta}
\alias{max_negdelta}
\title{max_negdelta}
\usage{
max_negdelta(x, end = 60)
}
\description{
Helper function for reordering levels of calcium response based on the maximum change in reponse
aligned to the end of a pulse, in downward responses. Mainly for use in plotGCaMP_multi.R. End value is derived from input
in plotGCaMP_multi.R
}
\examples{
ggplot(aes(x = time, y = fct_reorder(animal_num, delF, max_delta)))
}
