% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_matData.R
\name{melt_matData}
\alias{melt_matData}
\title{melt_matData
fucntion to tidy up matlab datasets which are in long format and separated by assay}
\usage{
melt_matData(df, datatype)
}
\arguments{
\item{df}{input data list, needs to be from readMat input, with naemed conditions}

\item{datatype}{type of data you are reorganizing (ie state, pulse # etc..)}
}
\description{
melt_matData
fucntion to tidy up matlab datasets which are in long format and separated by assay
}
\examples{
melt_matData(states, datatype = state)
}
