% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_FIJI_data.R
\name{merge_FIJI_data}
\alias{merge_FIJI_data}
\title{merge_FIJI_data}
\usage{
merge_FIJI_data(neuronfile, backsub = TRUE, show.plots = TRUE,
  frame.rate = 4)
}
\arguments{
\item{filename}{}
}
\description{
This helper function takes a group of measurements output from registerquantGCaMP.ijm imageJ macro
and generates a merged .csv file with mean neuron pixel values, mean background pixel values, and the
background-subtracted pixel values - ie the 'signal'. No bleaching correction is applied. Files are written
in place to the enclosing folder.
}
