% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_beh_enter.R
\name{norm_beh_enter}
\alias{norm_beh_enter}
\title{norm_beh_enter function
Normalizes behavioral data +/- 2.5mm for animals entering a
single stripe of chemotaxis cue.}
\usage{
norm_beh_enter(df)
}
\arguments{
\item{df}{input merged data.frame containing output from beh_mat, dir_mat and y_mat files.}
}
\description{
norm_beh_enter function
Normalizes behavioral data +/- 2.5mm for animals entering a
single stripe of chemotaxis cue.
}
\examples{
norm_beh_enter()
}

