% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_beh_exit_3stripe.R
\name{norm_beh_exit_3}
\alias{norm_beh_exit_3}
\title{norm_beh_exit_3 function
Normalizes behavioral data +/- 2.5mm for animals EXITING
3 stripes of chemotaxis cue.}
\usage{
norm_beh_exit_3(df)
}
\arguments{
\item{df}{input merged data.frame containing output from beh_mat, dir_mat and y_mat files.}
}
\description{
norm_beh_exit_3 function
Normalizes behavioral data +/- 2.5mm for animals EXITING
3 stripes of chemotaxis cue.
}
\examples{
norm.exit<-norm_beh_exit_3(df)
}
