% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap
Generate a heatmap from a previously analyzed GCaMP dataset using plotGCaMP_multi - output is ggplot object which can be modified further}
\usage{
plot_heatmap(heatmap_limits = "auto", response_time = 59.5, ...)
}
\arguments{
\item{heatmap_limits}{optional 3-value vector defining the color scale limits, ie c(-1,0,2)}

\item{response_time}{time of expected GCaMP response in seconds. heatmaps will be arranged in descending amplitude based on these responses.}
}
\description{
plot_heatmap
Generate a heatmap from a previously analyzed GCaMP dataset using plotGCaMP_multi - output is ggplot object which can be modified further
}
\examples{
plot <- plot_heatmap()

}
