% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Residency.R
\name{plot_Residency}
\alias{plot_Residency}
\title{Plot residency from Wormlab dauer data by space and time.}
\usage{
plot_Residency(df, time_bin, y_bin, nopause)
}
\arguments{
\item{df}{data frame}

\item{time_bin}{number of bins to split plot}

\item{y_bin}{number of position bins}
}
\description{
Plot residency from Wormlab dauer data by space and time.
}
\examples{
read.csv("file") \%>\% plot_Residency(time_bine = 4, y_bin = 100, nopause = TRUE)

}
