% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_speedFood.R
\name{plot_speedFood}
\alias{plot_speedFood}
\title{plot_speedFood
plot roaming speed across food conditions}
\usage{
plot_speedFood(...)
}
\description{
plot_speedFood
plot roaming speed across food conditions
}
\examples{
plot_speedFood()
}
