% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stateProbs.R
\name{plot_stateProbs}
\alias{plot_stateProbs}
\title{Plot state probability from Wormlab data by space and time.}
\usage{
plot_stateProbs(time_bin, y_bin)
}
\arguments{
\item{time_bin}{number of bins to split plot}

\item{y_bin}{number of position bins}

\item{df}{data frame}
}
\description{
Plot state probability from Wormlab data by space and time.
}
\examples{
plot_stateProbs <-

}
