% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAvgResidency.R
\name{plotAvgResidency}
\alias{plotAvgResidency}
\title{plotAvgResidency}
\usage{
plotAvgResidency(
  folderPath,
  y_bins = 50,
  y_max = 4,
  fillcolor = "grey",
  bordercolor = NA,
  ...
)
}
\arguments{
\item{FileFilter}{string to search/subset filenames}
}
\description{
Plots average residency for stripe data
}
\examples{
data <- plotAvgResidency()
}
