% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGCaMP_multi.R
\name{plotGCaMP_multi}
\alias{plotGCaMP_multi}
\title{plotGCaMP_multi}
\usage{
plotGCaMP_multi(FileFilter, genotype, cue = cue, food = OP50, start = 30,
  end = 60, center_on_pulse = FALSE, ...)
}
\arguments{
\item{FileFilter}{string to search/subset filenames}

\item{genotype}{label the genotype for these data}

\item{cue}{label the stimulus cue.}

\item{food}{label to food cue}

\item{start}{begin of stimulus}

\item{end}{end time of stimulus}

\item{center_on_pulse}{optional parameter to center delF values by the mean of the stimulus duration}
}
\description{
Function is a wrapper for exp.fit.all.log.lin which outputs corrected GCaMP signals as well as plots showing
original and corrected signals. Also plots average trace for GCaMP signal. Inputs are matfiles.
The script searches recursively for matfiles from a starter file (can be any placeholder file). Need to exclude
matfiles that are not GCaMP files either using FileFilter, or by putting only relevant files in the folder.
Requires max_delta helper function
time etc...
}
\examples{
data <- plotGCaMP_multi(N2, genotype = N2, cue = octanol)

}
