% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prism_scatter.R
\name{Prism_scatter}
\alias{Prism_scatter}
\title{Prism_scatter}
\usage{
Prism_scatter(data, xval, yval, facet, color_var, shape, ncolors, nbreaks, ...)
}
\arguments{
\item{data}{input dataset}

\item{xval}{main x-value (can be interaction of two factors)}

\item{yval}{y-values (continuous data)}

\item{facet}{factor to facet data}

\item{shape}{factor to set shape of scatterplot}

\item{color}{factor to set colors}

\item{ncolor}{number of colors to include (must match )}
}
\description{
This function replicates prism-like scatter plots with a bar-graph and a scatterplot with SEM values.
It requires discrete x, but can use interaction of multiple values to set x. Default is to colour value #1 black, and all else blue.
output is a ggplot object that can be further modified.
}
\examples{
p <- Prism_scatter(data, x = interaction(treatment,stage), y = odr10GFP, color = Dev_experience, shape = treatment, facet = Dev_experience, ncolors = 2)
}
