% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_binsize.R
\name{read.binsize}
\alias{read.binsize}
\title{read_binsize function
simple prompt for binsize - user input}
\usage{
read.binsize()
}
\arguments{
\item{n}{empty input - will take user input}
}
\description{
read_binsize function
simple prompt for binsize - user input
}
\examples{
n <- read_binsize()
}
