% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_movie.R
\name{register_movie}
\alias{register_movie}
\title{register_movie}
\usage{
register_movie()
}
\description{
simple R function to run an imageJ registration macro within R. Can be used to loop
through a list of video files.
}
