% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd_mat_tidy.R
\name{spd_mat_tidy}
\alias{spd_mat_tidy}
\title{spd_mat_tidy data munge
Takes a matlab ymat file and converts it to long format for further analysis}
\usage{
spd_mat_tidy()
}
\arguments{
\item{bins}{number of time frames to bin. generally 400 for 20 minute video is standard}
}
\description{
spd_mat_tidy data munge
Takes a matlab ymat file and converts it to long format for further analysis
}
\examples{
spd_mat_tidy()
}
