% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Speed_curve_4xParallel.R
\name{Speed_curve_4xParallel}
\alias{Speed_curve_4xParallel}
\title{Speed_curve_4xParallel
roam/dwell test using wormlab data}
\usage{
Speed_curve_4xParallel(bin.length = 10, frame.rate, ...)
}
\arguments{
\item{bin.length}{length of time bins in seconds. Used for state analysis}

\item{frame.rate}{video frame rate}

\item{num.tracks}{optional argument to limit input to certain number of worm tracks}
}
\description{
Speed_curve_4xParallel
roam/dwell test using wormlab data
}
\examples{
Speed_curve_4xParallel(bin.length = 10, ...)
}
