% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLmult_dauer_res.R
\name{WLmult_dauer_res}
\alias{WLmult_dauer_res}
\title{Wrapper for 'Import.WL.data' and 'plot_Residency' to allow multiple datasets to be simultaneously
analyzed. Uses recursive search for a *position.csv file, then makes a file list.}
\usage{
WLmult_dauer_res()
}
\arguments{
\item{bin.length}{length of time bins in seconds. Used for state analysis}

\item{frame.rate}{video frame rate}

\item{num.tracks}{optional argument to limit input to certain number of worm tracks}
}
\description{
Wrapper for 'Import.WL.data' and 'plot_Residency' to allow multiple datasets to be simultaneously
analyzed. Uses recursive search for a *position.csv file, then makes a file list.
}
\examples{
WLmult_dauer_res()
}
