% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteStripeProp.R
\name{WriteStripeProp}
\alias{WriteStripeProp}
\title{using all_track_data.csv file, calculates proportion of worm tracks in (1) or out (0) of a stripe
uses GetStripeProp to determine whether tracks are in or out
Writes the proportion data to a csv within the same folder.}
\usage{
WriteStripeProp(...)
}
\description{
using all_track_data.csv file, calculates proportion of worm tracks in (1) or out (0) of a stripe
uses GetStripeProp to determine whether tracks are in or out
Writes the proportion data to a csv within the same folder.
}
\examples{
WriteStripeProp()
}
