% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Y_mat.R
\name{y_mat}
\alias{y_mat}
\title{y_mat melt function
Takes a matlab ymat file and converts it to long format for further analysis}
\usage{
y_mat(n)
}
\arguments{
\item{n}{number of time frames to bin. generally 400 for 20 minute video is standard}
}
\description{
y_mat melt function
Takes a matlab ymat file and converts it to long format for further analysis
}
\examples{
y_mat()
}
