% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_mat_tidy.R
\name{y_mat_tidy}
\alias{y_mat_tidy}
\title{y_mat_tidy data munge
Takes a matlab ymat file and converts it to long format for further analysis}
\usage{
y_mat_tidy()
}
\arguments{
\item{bins}{number of time frames to bin. generally 400 for 20 minute video is standard}
}
\description{
y_mat_tidy data munge
Takes a matlab ymat file and converts it to long format for further analysis
}
\examples{
y_mat_tidy()
}
