% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.fit.all.log.lin.R
\name{exp.fit.all.log.lin}
\alias{exp.fit.all.log.lin}
\title{exp.fit.all.log.lin}
\usage{
\method{exp}{fit.all.log.lin}(filename, skip.time, matlab = TRUE,
  show.plots = TRUE, linear = TRUE, nls = FALSE, startPulse = 29.5,
  endPulse = 60.5)
}
\arguments{
\item{filename}{filepaths of .mat files which have a "signal"  and "time" field.}

\item{skip.time}{number of seconds to skip at the beginning for the exponential fit. N=10 improves the fit.}

\item{nls}{use nls to perform exponential fit: signal ~ y0 + exp((-time - T0) / tau)}
}
\description{
Function takes a list of calcium imaging files by filenames in the current working directory (.mat files)
and performs a simple exponential regression fit to normalize the deltaF/F signal. This function incorporates a linear
to account for gradual rise in signal, which can alter the single exponential. Will output plots showing orignal
and fitted values, as well as corrected values. If an object is assigned, this will be a vector of corrected values
}
\examples{
data <- exp.fit.all.log.lin(files[1], skip.time = 10)

}
