% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Residency.R
\name{plot_Residency}
\alias{plot_Residency}
\title{Plot residency from Wormlab dauer data by space and time.}
\usage{
plot_Residency(time_bin, y_bin, overlay, multiple, stripe.index = FALSE, ...)
}
\arguments{
\item{time_bin}{number of bins to split plot}

\item{y_bin}{number of position bins}

\item{overlay}{optional argument to plot overlay of density for 2 time bins}
}
\description{
Plot residency from Wormlab dauer data by space and time.
}
\examples{
plot_Residency(time_bin = 4, y_bin = 100, nopause = TRUE)

}
