% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGCaMP_multi.R
\name{plotGCaMP_multi}
\alias{plotGCaMP_multi}
\title{plotGCaMP_multi}
\usage{
plotGCaMP_multi(FileFilter, matlab = FALSE, genotype, cue = cue,
  food = OP50, startPulse = 29.5, endPulse = 59.5,
  center_on_pulse = FALSE, show.plots = TRUE, use.Fmax = FALSE,
  neuron = GCAMP, linear = FALSE, nls = TRUE, backsub = FALSE,
  plot.raw = FALSE, heatmap_limits = "auto", folderPath, ...)
}
\arguments{
\item{FileFilter}{string to search/subset filenames}

\item{matlab}{whether to use matfiles or data from ImageJ quantification. Defaults to TRUE (matfiles)}

\item{genotype}{label the genotype for these data}

\item{cue}{label the stimulus cue.}

\item{food}{label to food cue}

\item{startPulse}{begin of stimulus}

\item{endPulse}{endPulse time of stimulus}

\item{center_on_pulse}{optional parameter to center delF values by the mean of the stimulus duration
'OFF' = Bring values to mean delF of 2nd half of pulse duration, order by OFF responses
'ON' = Bring values to mean delF of 2nd half pre-pulse duration, order by ON responses}

\item{show.plots}{render plots for baseline correction - defaults to TRUE}

\item{use.Fmax}{normalize amplitude to 1}

\item{neuron}{neuron being analyzed}

\item{linear}{optional argument piped into exp.fit.all.log.lin include a linear term in the fit?}

\item{backsub}{do you want to subtract background ie subtract ROI of background from the mean cell ROI? Defaults to TRUE}

\item{plot.raw}{do you want to use raw deltaF/F values - ie not exponentially corrected? Defaults to FALSE}

\item{heatmap_limits}{optional 3-value vector defining the color scale and y axis limits, ie c(-1,0,2)}

\item{folderpath}{user supplied path for recursive file search. If missing, it will prompt for a fiel selection.}
}
\description{
Function is a wrapper for exp.fit.all.log.lin which outputs corrected GCaMP signals as well as plots showing
original and corrected signals. Also plots average trace for GCaMP signal. Inputs are matfiles with a "signal" field,
or .csv files from imageJ analysis.
The script searches recursively for matfiles (*.mat) or csv files (\*.csv) from a file (can be any placeholder file). Need to exclude
matfiles that are not GCaMP files either using FileFilter, or by putting only relevant files in the folder. Default
startpulse assumes 400ms delay for camera recording.
Requires max_delta helper function
time etc...
}
\examples{
data <- plotGCaMP_multi(N2, genotype = N2, cue = octanol)

}
